traverse dependency graph to create subgraphs
