+++ src/utils/io.ts
@@ -1,0 +1,19 @@
+import { Readable } from 'stream'
+
+export async function streamToBuffer (stream: Readable): Promise<Buffer> {
+  return new Promise((resolve, reject) => {
+    const data: Uint8Array[] = []
+
+    stream.on('data', (chunk) => {
+      data.push(chunk)
+    })
+
+    stream.on('end', () => {
+      resolve(Buffer.concat(data))
+    })
+
+    stream.on('error', (err) => {
+      reject(err)
+    })
+  })
+}
